requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Storage
sqlalchemy==2.0.25
aiosqlite==0.19.0
chromadb==0.4.22

# AI Providers
httpx==0.26.0
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.18.1

# Adapters
python-telegram-bot==21.0.1
slack-sdk==3.27.1
aiosmtplib==3.0.1
aiofiles==23.2.1

# Utilities
pyyaml==6.0.1
python-multipart==0.0.9
jinja2==3.1.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
